requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-filter==2.4.0
pandas==1.2.4
django-storages==1.11.1
boto3==1.17.93
SQLAlchemy==1.4.18
